tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/115.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: Removing src of defer script before it runs </title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="test"></div>
        <div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script id="defer-script" defer src="scripts/include-1.js">t.step(function() {assert_unreached()})</script>

        <script>
          var t = async_test();

          document.getElementById("defer-script").removeAttribute("src");

          addEventListener("load", t.step_func(function () {
            assert_array_equals(eventOrder, ["external script #1"]);
            t.done();
          }), false);
        </script>

</body></html>
